Setting modhash to null; odd parameters
The bug I found previously had to do with the modhash being completely empty. My guess is when I was using an incorrect command somewhere. It is very unusual for the modhash to be set to null or empty, so I added a print statement for now, as well as a stack trace